Update to 1.18.2 (#268)
* port to 1.18.2

* add first gametest

* bump forge and potionbundles and fix min forge version

* make runtime only deps not required for ci

+++ src/main/java/com/github/minecraftschurlimods/arsmagicalegacy/compat/patchouli/CapStateMatcher.java
@@ -17,8 +17,8 @@ class CapStateMatcher implements IStateMatcher {
     }
 
     @Override
-    public BlockState getDisplayedState(int ticks) {
-        AltarCapMaterial mat = AltarMaterialManager.instance().getRandomCapMaterial(ticks / 20);
+    public BlockState getDisplayedState(long ticks) {
+        AltarCapMaterial mat = AltarMaterialManager.instance().getRandomCapMaterial(Math.toIntExact(ticks / 20));
         if (mat == null) return Blocks.AIR.defaultBlockState();
         return mat.cap().defaultBlockState();
     }
